+++ src/Elements/Spatial/MercatorProjection.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Elements.GeoJSON
+namespace Elements.Spatial
 {
     /// <summary>
     /// Methods for computing geographic coordinates using the Mercator projection.
@@ -13,19 +13,15 @@ namespace Elements.GeoJSON
         private static readonly double ECCENT = Math.Sqrt(1.0 - (RATIO * RATIO));
         private static readonly double COM = 0.5 * ECCENT;
 
-        private static readonly double DEG2RAD = Math.PI / 180.0;
-        private static readonly double RAD2Deg = 180.0 / Math.PI;
-        private static readonly double PI_2 = Math.PI / 2.0;
-
         /// <summary>
         /// Get the coordinates of the longitude and latitude.
         /// </summary>
         /// <param name="lon"></param>
         /// <param name="lat"></param>
         /// <returns>An array of doubles containing the x, and y coordintes, in meters.</returns>
-        public static double[] toPixel(double lon, double lat)
+        public static double[] ToPixel(double lon, double lat)
         {
-            return new double[] { lonToX(lon), latToY(lat) };
+            return new double[] { LonToX(lon), LatToY(lat) };
         }
         
         /// <summary>
@@ -34,9 +30,9 @@ namespace Elements.GeoJSON
         /// <param name="x"></param>
         /// <param name="y"></param>
         /// <returns>An array of doubles containing the longitude and latitude in degrees.</returns>
-        public static double[] toGeoCoord(double x, double y)
+        public static double[] ToGeoCoord(double x, double y)
         {
-            return new double[] { xToLon(x), yToLat(y) };
+            return new double[] { XToLon(x), YToLat(y) };
         }
 
         /// <summary>
@@ -44,9 +40,9 @@ namespace Elements.GeoJSON
         /// </summary>
         /// <param name="lon"></param>
         /// <returns></returns>
-        public static double lonToX(double lon)
+        public static double LonToX(double lon)
         {
-            return R_MAJOR * DegToRad(lon);
+            return R_MAJOR * Units.DegreesToRadians(lon);
         }
 
         /// <summary>
@@ -54,10 +50,10 @@ namespace Elements.GeoJSON
         /// </summary>
         /// <param name="lat"></param>
         /// <returns></returns>
-        public static double latToY(double lat)
+        public static double LatToY(double lat)
         {
             lat = Math.Min(89.5, Math.Max(lat, -89.5));
-            double phi = DegToRad(lat);
+            double phi = Units.DegreesToRadians(lat);
             double sinphi = Math.Sin(phi);
             double con = ECCENT * sinphi;
             con = Math.Pow(((1.0 - con) / (1.0 + con)), COM);
@@ -70,9 +66,9 @@ namespace Elements.GeoJSON
         /// </summary>
         /// <param name="x">The x coordinate.</param>
         /// <returns>The longitude in degrees.</returns>
-        public static double xToLon(double x)
+        public static double XToLon(double x)
         {
-            return RadToDeg(x) / R_MAJOR;
+            return Units.RadiansToDegrees(x) / R_MAJOR;
         }
 
         /// <summary>
@@ -80,30 +76,20 @@ namespace Elements.GeoJSON
         /// </summary>
         /// <param name="y">The y coordinate.</param>
         /// <returns>The latitude in degrees.</returns>
-        public static double yToLat(double y)
+        public static double YToLat(double y)
         {
             double ts = Math.Exp(-y / R_MAJOR);
-            double phi = PI_2 - 2 * Math.Atan(ts);
+            double phi = Units.PI_2 - 2 * Math.Atan(ts);
             double dphi = 1.0;
             int i = 0;
             while ((Math.Abs(dphi) > 0.000000001) && (i < 15))
             {
                 double con = ECCENT * Math.Sin(phi);
-                dphi = PI_2 - 2 * Math.Atan(ts * Math.Pow((1.0 - con) / (1.0 + con), COM)) - phi;
+                dphi = Units.PI_2 - 2 * Math.Atan(ts * Math.Pow((1.0 - con) / (1.0 + con), COM)) - phi;
                 phi += dphi;
                 i++;
             }
-            return RadToDeg(phi);
-        }
-
-        private static double RadToDeg(double rad)
-        {
-            return rad * RAD2Deg;
-        }
-
-        private static double DegToRad(double deg)
-        {
-            return deg * DEG2RAD;
+            return Units.RadiansToDegrees(phi);
         }
     }
 }