+++ test/Elements.Tests/Examples/InstanceExample.cs
@@ -2,8 +2,8 @@ using Elements.Geometry;
 using Xunit;
 
 namespace Elements.Tests.Examples
- {
-     public class InstanceExample : ModelTest
+{
+    public class InstanceExample : ModelTest
      {
         [Fact]
         public void Instance()
@@ -11,19 +11,20 @@ namespace Elements.Tests.Examples
             this.Name = "Elements_Instance";
 
             // <example>
-            var mass = new Mass(Polygon.Rectangle(1.0, 1.0),1.0, new Material("yellow", Colors.Yellow));
-            this.Model.AddElement(mass);
+            // Create a mass type from which to create instances.
+            var m = new Mass(new Profile(Polygon.Rectangle(1.0, 1.0)), 1.5);
 
             var j = 1.0;
-            var count = 10;
+            var count = 20;
             for(var i=0.0; i<360.0; i+= 360.0/(double)count)
             {
-                var m2 = new Instance(mass);
-                m2.Transform.Scale(new Vector3(j,j,j));
-                m2.Transform.Move(new Vector3(3, 0, 0));
-                m2.Transform.Rotate(Vector3.ZAxis, i);
-                this.Model.AddElement(m2);
-
+                var t = new Transform();
+                t.Scale(new Vector3(j,j,j));
+                t.Move(new Vector3(3, 0, 0));
+                t.Rotate(Vector3.ZAxis, i);
+                var instance = m.CreateInstance(t, $"mass {i}");
+                this.Model.AddElement(instance);
+                
                 j += 1.0/(double)count;
             }
             //</example>